requirements.txt
pillow=11.3.0
pillow_heif=1.3.0
pillow-avif-plugin=1.5.5
pymupdf=1.27.2.2